backend/requirements.txt
# Todo Intelligence Platform - Backend Dependencies
# Python 3.11+

# ==============================================
# Core Framework
# ==============================================
fastapi[all]>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ==============================================
# Database
# ==============================================
sqlalchemy[asyncio]>=2.0.0
sqlmodel>=0.0.14
aiosqlite>=0.19.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# ==============================================
# Migrations
# ==============================================
alembic>=1.12.0

# ==============================================
# Authentication
# ==============================================
# better-auth (to be added in Stage 7)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# ==============================================
# Environment & Configuration
# ==============================================
python-dotenv>=1.0.0

# ==============================================
# HTTP Client
# ==============================================
httpx>=0.25.0

# ==============================================
# AI Integration
# ==============================================
groq>=0.4.0

# ==============================================
# Date/Time Handling
# ==============================================
python-dateutil>=2.8.2

# ==============================================
# Testing
# ==============================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0  # For testing FastAPI

# ==============================================
# Code Quality
# ==============================================
black>=23.0.0
pylint>=3.0.0
mypy>=1.7.0

# ==============================================
# Monitoring & Logging (Stage 8)
# ==============================================
# sentry-sdk>=1.38.0

# ==============================================
# Rate Limiting (Stage 7)
# ==============================================
# slowapi>=0.1.9
